Выставил цель для игроков

+++ Assets/Scripts/FusionCore/Ui/IndicatorView.cs
@@ -11,7 +11,7 @@ namespace FusionCore.Ui
         [SerializeField] 
         private TMP_Text _countArmor;
 
-        public void SetData(int countHealth, int countArmor)
+        public void SetData(float countHealth, float countArmor)
         {
             _countHealth.text = $"{countHealth} health";
             _countArmor.text = $"{countArmor} armor";
